Incluindo sistema de pagamento (pix) e mudando os models.

+++ Commands/AddProductCommand.cs
@@ -27,7 +27,7 @@ namespace bot.Commands
                 await BotClient.SendMessage(Update.Message.Chat.Id, "⚠️ Você precisa informar os dados do produto. Use este exemplo:\n\n<pre>Nome|descrição|preço|imagem|categoria</pre>", parseMode: Telegram.Bot.Types.Enums.ParseMode.Html);
                 return;
             }
-            Product newProduct = new() { Name = msg[0].Split()[1], Description = msg[1], Price = double.Parse(msg[2]), ImageUrl = msg[3], Category = msg[4] };
+            Product newProduct = new() { Name = msg[0].Split()[1], Description = msg[1], Price = decimal.Parse(msg[2]), ImageUrl = msg[3], Category = msg[4] };
             await _shop.AddProduct(newProduct);
             await BotClient.SendMessage(Update.Message.Chat.Id, "O Produto foi adicionado.");
         }
